requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
python-multipart
pyjwt
bcrypt
aiofiles
motor
pymongo
mcp[cli]
